requirements.txt
opencv-python
pyautogui
mss
numpy
Pillow
asyncio
tkinter
